requirements.txt
fiftyone>=0.23.0

numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0

PyYAML>=6.0
hydra-core>=1.3.0

tqdm>=4.65.0

ujson>=5.8.0

ultralytics>=8.0.0
onnxruntime-gpu==1.20.0

anywidget>=0.9.0
